Refactor to allow use in parameters
The puppet-boolean functionality is incredibly helpful as it exists
currently. However, it is locked down to be used only in property based
attributes.

This commit moves the shared behavior into a module in the `PuppetX`
space. This module is then included by both 'Puppet::Parameter::Boolean'
and `Puppet::Property::Boolean` so that this functionality may be made
available to both kinds of attributes.